+++ bundle/src/test/java/com/adobe/acs/commons/http/headers/AbstractDispatcherCacheHeaderFilterTest.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package com.adobe.acs.commons.wcm.filter.impl;
+package com.adobe.acs.commons.http.headers;
 
 import static org.mockito.Matchers.*;
 import static org.mockito.Mockito.*;
@@ -50,6 +50,8 @@ import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.cm.ConfigurationException;
 import org.osgi.service.component.ComponentContext;
+
+import com.adobe.acs.commons.http.headers.AbstractDispatcherCacheHeaderFilter;
 
 @RunWith(MockitoJUnitRunner.class)
 public class AbstractDispatcherCacheHeaderFilterTest {
