+++ bundle/src/test/java/com/adobe/acs/commons/http/headers/AbstractExpiresHeaderFilterTest.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package com.adobe.acs.commons.wcm.filter.impl;
+package com.adobe.acs.commons.http.headers;
 
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
@@ -44,6 +44,9 @@ import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.osgi.service.cm.ConfigurationException;
 import org.osgi.service.component.ComponentContext;
+
+import com.adobe.acs.commons.http.headers.AbstractDispatcherCacheHeaderFilter;
+import com.adobe.acs.commons.http.headers.AbstractExpiresHeaderFilter;
 
 @RunWith(MockitoJUnitRunner.class)
 public class AbstractExpiresHeaderFilterTest {
