+++ bundle/src/test/java/com/adobe/acs/commons/http/headers/MonthlyExpiresHeaderFilterTest.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package com.adobe.acs.commons.wcm.filter.impl;
+package com.adobe.acs.commons.http.headers;
 
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
@@ -37,10 +37,12 @@ import org.osgi.framework.BundleContext;
 import org.osgi.service.cm.ConfigurationException;
 import org.osgi.service.component.ComponentContext;
 
-@RunWith(MockitoJUnitRunner.class)
-public class WeeklyExpiresHeaderFilterTest {
+import com.adobe.acs.commons.http.headers.MonthlyExpiresHeaderFilter;
 
-    WeeklyExpiresHeaderFilter filter = new WeeklyExpiresHeaderFilter();
+@RunWith(MockitoJUnitRunner.class)
+public class MonthlyExpiresHeaderFilterTest {
+
+    MonthlyExpiresHeaderFilter filter = new MonthlyExpiresHeaderFilter();
 
     Dictionary<String, Object> properties = null;
 
@@ -53,7 +55,7 @@ public class WeeklyExpiresHeaderFilterTest {
     @Before
     public void setup() throws Exception {
         properties = new Hashtable<String, Object>();
-        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_TIME, "02:30");
+        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_TIME, "02:30");
 
         when(componentContext.getProperties()).thenReturn(properties);
         when(componentContext.getBundleContext()).thenReturn(bundleContext);
@@ -67,8 +69,7 @@ public class WeeklyExpiresHeaderFilterTest {
     }
 
     @Test
-    public void testAdjustExpiresPastWeekday() throws Exception {
-
+    public void testAdjustExpiresPastDay() throws Exception {
 
         Calendar actual = Calendar.getInstance();
         actual.set(Calendar.SECOND, 0);
@@ -76,44 +77,44 @@ public class WeeklyExpiresHeaderFilterTest {
 
         Calendar expected = Calendar.getInstance();
         expected.setTime(actual.getTime());
-        expected.add(Calendar.DAY_OF_WEEK, -3);
-        expected.add(Calendar.DAY_OF_WEEK, 7);
+        expected.add(Calendar.DAY_OF_MONTH, -3);
+        expected.add(Calendar.MONTH, 1);
 
-        int dayOfWeek = expected.get(Calendar.DAY_OF_WEEK);
-        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, dayOfWeek);
+        actual.set(Calendar.DAY_OF_MONTH, 15);
+
+        int month = expected.get(Calendar.MONTH);
+        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, expected.get(Calendar.DAY_OF_MONTH));
 
         filter.doActivate(componentContext);
         filter.adjustExpires(actual);
 
         assertTrue(DateUtils.isSameInstant(expected, actual));
-        assertEquals(dayOfWeek, actual.get(Calendar.DAY_OF_WEEK));
+        assertEquals(month, actual.get(Calendar.MONTH));
     }
 
     @Test
     public void testAdjustExpiresSameDayPast() throws Exception {
 
-
         Calendar actual = Calendar.getInstance();
         actual.set(Calendar.SECOND, 0);
         actual.set(Calendar.MILLISECOND, 0);
 
         Calendar expected = Calendar.getInstance();
         expected.setTime(actual.getTime());
-        expected.add(Calendar.DAY_OF_WEEK, 7);
+        expected.add(Calendar.MONTH, 1);
 
-        int dayOfWeek = expected.get(Calendar.DAY_OF_WEEK);
-        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, dayOfWeek);
+        int month = expected.get(Calendar.MONTH);
+        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, expected.get(Calendar.DAY_OF_MONTH));
 
         filter.doActivate(componentContext);
         filter.adjustExpires(actual);
 
         assertTrue(DateUtils.isSameInstant(expected, actual));
-        assertEquals(dayOfWeek, actual.get(Calendar.DAY_OF_WEEK));
+        assertEquals(month, actual.get(Calendar.MONTH));
     }
 
     @Test
     public void testAdjustExpiresSameDayFuture() throws Exception {
-
 
         Calendar actual = Calendar.getInstance();
         actual.add(Calendar.MINUTE, 1);
@@ -122,19 +123,19 @@ public class WeeklyExpiresHeaderFilterTest {
 
         Calendar expected = Calendar.getInstance();
         expected.setTime(actual.getTime());
-        int dayOfWeek = expected.get(Calendar.DAY_OF_WEEK);
-        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, dayOfWeek);
+
+        int month = expected.get(Calendar.MONTH);
+        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, expected.get(Calendar.DAY_OF_MONTH));
 
         filter.doActivate(componentContext);
         filter.adjustExpires(actual);
 
         assertTrue(DateUtils.isSameInstant(expected, actual));
-        assertEquals(dayOfWeek, actual.get(Calendar.DAY_OF_WEEK));
+        assertEquals(month, actual.get(Calendar.MONTH));
     }
 
     @Test
-    public void testAdjustExpiresFutureWeekday() throws Exception {
-
+    public void testAdjustExpiresFutureDay() throws Exception {
 
         Calendar actual = Calendar.getInstance();
         actual.set(Calendar.SECOND, 0);
@@ -142,24 +143,69 @@ public class WeeklyExpiresHeaderFilterTest {
 
         Calendar expected = Calendar.getInstance();
         expected.setTime(actual.getTime());
-        expected.add(Calendar.DAY_OF_WEEK, 2);
+        expected.add(Calendar.DAY_OF_MONTH, 5);
 
-        int dayOfweek = expected.get(Calendar.DAY_OF_WEEK);
-        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, dayOfweek);
+        actual.set(Calendar.DAY_OF_MONTH, 15);
 
+        int month = expected.get(Calendar.MONTH);
+        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, expected.get(Calendar.DAY_OF_MONTH));
 
         filter.doActivate(componentContext);
         filter.adjustExpires(actual);
 
         assertTrue(DateUtils.isSameInstant(expected, actual));
-        assertEquals(dayOfweek, actual.get(Calendar.DAY_OF_WEEK));
+        assertEquals(month, actual.get(Calendar.MONTH));
     }
 
+    @Test
+    public void testAdjustExpiresLastDayPast() throws Exception {
+
+        Calendar actual = Calendar.getInstance();
+        actual.add(Calendar.MONTH, -1);
+        actual.set(Calendar.SECOND, 0);
+        actual.set(Calendar.MILLISECOND, 0);
+        actual.set(Calendar.DAY_OF_MONTH, actual.getActualMaximum(Calendar.DAY_OF_MONTH));
+
+        Calendar expected = Calendar.getInstance();
+        expected.setTime(actual.getTime());
+        expected.add(Calendar.MONTH, 1);
+
+        int month = expected.get(Calendar.MONTH);
+        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, "LAST");
+
+        filter.doActivate(componentContext);
+        filter.adjustExpires(actual);
+
+        assertTrue(DateUtils.isSameInstant(expected, actual));
+        assertEquals(month, actual.get(Calendar.MONTH));
+    }
+
+    @Test
+    public void testAdjustExpiresLastDayFuture() throws Exception {
+
+        Calendar actual = Calendar.getInstance();
+        actual.add(Calendar.MINUTE, 1);
+        actual.set(Calendar.SECOND, 0);
+        actual.set(Calendar.MILLISECOND, 0);
+        actual.set(Calendar.DAY_OF_MONTH, actual.getActualMaximum(Calendar.DAY_OF_MONTH));
+
+        Calendar expected = Calendar.getInstance();
+        expected.setTime(actual.getTime());
+
+        int month = expected.get(Calendar.MONTH);
+        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, "LAST");
+
+        filter.doActivate(componentContext);
+        filter.adjustExpires(actual);
+
+        assertTrue(DateUtils.isSameInstant(expected, actual));
+        assertEquals(month, actual.get(Calendar.MONTH));
+    }
 
     @Test
     public void testDoActivateSuccess() throws Exception {
 
-        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, Calendar.SUNDAY);
+        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, 15);
 
         Calendar actual = Calendar.getInstance();
         actual.set(Calendar.SECOND, 0);
@@ -168,8 +214,9 @@ public class WeeklyExpiresHeaderFilterTest {
         filter.doActivate(componentContext);
         filter.adjustExpires(actual);
 
-        assertEquals(Calendar.SUNDAY, actual.get(Calendar.DAY_OF_WEEK));
-        // Abstract class will pass expires, this checks we aren't messing with the time.
+        assertEquals(15, actual.get(Calendar.DAY_OF_MONTH));
+        // Abstract class will pass expires, this checks we aren't messing with
+        // the time.
         assertNotEquals(2, actual.get(Calendar.HOUR_OF_DAY));
         assertNotEquals(30, actual.get(Calendar.MINUTE));
     }
@@ -177,25 +224,27 @@ public class WeeklyExpiresHeaderFilterTest {
     @Test(expected = ConfigurationException.class)
     public void testDoActivateCallsParent() throws Exception {
 
-        properties.remove(WeeklyExpiresHeaderFilter.PROP_EXPIRES_TIME);
+        properties.remove(MonthlyExpiresHeaderFilter.PROP_EXPIRES_TIME);
         filter.doActivate(componentContext);
     }
 
     @Test(expected = ConfigurationException.class)
-    public void testDoActivateNoDayOfWeek() throws Exception {
+    public void testDoActivateNoDayOfMonth() throws Exception {
 
         filter.doActivate(componentContext);
     }
 
     @Test(expected = ConfigurationException.class)
-    public void testDoActivateInvalidLowDayOfWeek() throws Exception {
-        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, Calendar.SUNDAY - 1);
+    public void testDoActivateInvalidLowDayOfMonth() throws Exception {
+        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, 0);
         filter.doActivate(componentContext);
     }
 
     @Test(expected = ConfigurationException.class)
-    public void testDoActivateInvalidHighDayOfWeek() throws Exception {
-        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, Calendar.SATURDAY + 1);
+    public void testDoActivateInvalidHighDayOfMonth() throws Exception {
+        Calendar test = Calendar.getInstance();
+        int val = test.getActualMaximum(Calendar.DAY_OF_MONTH) + 1;
+        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, val);
         filter.doActivate(componentContext);
     }
 }
