+++ bundle/src/test/java/com/adobe/acs/commons/http/headers/WeeklyExpiresHeaderFilterTest.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package com.adobe.acs.commons.wcm.filter.impl;
+package com.adobe.acs.commons.http.headers;
 
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
@@ -37,10 +37,12 @@ import org.osgi.framework.BundleContext;
 import org.osgi.service.cm.ConfigurationException;
 import org.osgi.service.component.ComponentContext;
 
-@RunWith(MockitoJUnitRunner.class)
-public class MonthlyExpiresHeaderFilterTest {
+import com.adobe.acs.commons.http.headers.WeeklyExpiresHeaderFilter;
 
-    MonthlyExpiresHeaderFilter filter = new MonthlyExpiresHeaderFilter();
+@RunWith(MockitoJUnitRunner.class)
+public class WeeklyExpiresHeaderFilterTest {
+
+    WeeklyExpiresHeaderFilter filter = new WeeklyExpiresHeaderFilter();
 
     Dictionary<String, Object> properties = null;
 
@@ -53,7 +55,7 @@ public class MonthlyExpiresHeaderFilterTest {
     @Before
     public void setup() throws Exception {
         properties = new Hashtable<String, Object>();
-        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_TIME, "02:30");
+        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_TIME, "02:30");
 
         when(componentContext.getProperties()).thenReturn(properties);
         when(componentContext.getBundleContext()).thenReturn(bundleContext);
@@ -67,7 +69,8 @@ public class MonthlyExpiresHeaderFilterTest {
     }
 
     @Test
-    public void testAdjustExpiresPastDay() throws Exception {
+    public void testAdjustExpiresPastWeekday() throws Exception {
+
 
         Calendar actual = Calendar.getInstance();
         actual.set(Calendar.SECOND, 0);
@@ -75,135 +78,90 @@ public class MonthlyExpiresHeaderFilterTest {
 
         Calendar expected = Calendar.getInstance();
         expected.setTime(actual.getTime());
-        expected.add(Calendar.DAY_OF_MONTH, -3);
-        expected.add(Calendar.MONTH, 1);
+        expected.add(Calendar.DAY_OF_WEEK, -3);
+        expected.add(Calendar.DAY_OF_WEEK, 7);
 
-        actual.set(Calendar.DAY_OF_MONTH, 15);
-
-        int month = expected.get(Calendar.MONTH);
-        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, expected.get(Calendar.DAY_OF_MONTH));
+        int dayOfWeek = expected.get(Calendar.DAY_OF_WEEK);
+        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, dayOfWeek);
 
         filter.doActivate(componentContext);
         filter.adjustExpires(actual);
 
         assertTrue(DateUtils.isSameInstant(expected, actual));
-        assertEquals(month, actual.get(Calendar.MONTH));
+        assertEquals(dayOfWeek, actual.get(Calendar.DAY_OF_WEEK));
     }
 
     @Test
     public void testAdjustExpiresSameDayPast() throws Exception {
 
+
         Calendar actual = Calendar.getInstance();
         actual.set(Calendar.SECOND, 0);
         actual.set(Calendar.MILLISECOND, 0);
 
         Calendar expected = Calendar.getInstance();
         expected.setTime(actual.getTime());
-        expected.add(Calendar.MONTH, 1);
+        expected.add(Calendar.DAY_OF_WEEK, 7);
 
-        int month = expected.get(Calendar.MONTH);
-        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, expected.get(Calendar.DAY_OF_MONTH));
+        int dayOfWeek = expected.get(Calendar.DAY_OF_WEEK);
+        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, dayOfWeek);
 
         filter.doActivate(componentContext);
         filter.adjustExpires(actual);
 
         assertTrue(DateUtils.isSameInstant(expected, actual));
-        assertEquals(month, actual.get(Calendar.MONTH));
+        assertEquals(dayOfWeek, actual.get(Calendar.DAY_OF_WEEK));
     }
 
     @Test
     public void testAdjustExpiresSameDayFuture() throws Exception {
 
-        Calendar actual = Calendar.getInstance();
-        actual.add(Calendar.MINUTE, 1);
-        actual.set(Calendar.SECOND, 0);
-        actual.set(Calendar.MILLISECOND, 0);
-
-        Calendar expected = Calendar.getInstance();
-        expected.setTime(actual.getTime());
-
-        int month = expected.get(Calendar.MONTH);
-        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, expected.get(Calendar.DAY_OF_MONTH));
-
-        filter.doActivate(componentContext);
-        filter.adjustExpires(actual);
-
-        assertTrue(DateUtils.isSameInstant(expected, actual));
-        assertEquals(month, actual.get(Calendar.MONTH));
-    }
-
-    @Test
-    public void testAdjustExpiresFutureDay() throws Exception {
-
-        Calendar actual = Calendar.getInstance();
-        actual.set(Calendar.SECOND, 0);
-        actual.set(Calendar.MILLISECOND, 0);
-
-        Calendar expected = Calendar.getInstance();
-        expected.setTime(actual.getTime());
-        expected.add(Calendar.DAY_OF_MONTH, 5);
-
-        actual.set(Calendar.DAY_OF_MONTH, 15);
-
-        int month = expected.get(Calendar.MONTH);
-        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, expected.get(Calendar.DAY_OF_MONTH));
-
-        filter.doActivate(componentContext);
-        filter.adjustExpires(actual);
-
-        assertTrue(DateUtils.isSameInstant(expected, actual));
-        assertEquals(month, actual.get(Calendar.MONTH));
-    }
-
-    @Test
-    public void testAdjustExpiresLastDayPast() throws Exception {
-
-        Calendar actual = Calendar.getInstance();
-        actual.add(Calendar.MONTH, -1);
-        actual.set(Calendar.SECOND, 0);
-        actual.set(Calendar.MILLISECOND, 0);
-        actual.set(Calendar.DAY_OF_MONTH, actual.getActualMaximum(Calendar.DAY_OF_MONTH));
-
-        Calendar expected = Calendar.getInstance();
-        expected.setTime(actual.getTime());
-        expected.add(Calendar.MONTH, 1);
-
-        int month = expected.get(Calendar.MONTH);
-        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, "LAST");
-
-        filter.doActivate(componentContext);
-        filter.adjustExpires(actual);
-
-        assertTrue(DateUtils.isSameInstant(expected, actual));
-        assertEquals(month, actual.get(Calendar.MONTH));
-    }
-
-    @Test
-    public void testAdjustExpiresLastDayFuture() throws Exception {
 
         Calendar actual = Calendar.getInstance();
         actual.add(Calendar.MINUTE, 1);
         actual.set(Calendar.SECOND, 0);
         actual.set(Calendar.MILLISECOND, 0);
-        actual.set(Calendar.DAY_OF_MONTH, actual.getActualMaximum(Calendar.DAY_OF_MONTH));
 
         Calendar expected = Calendar.getInstance();
         expected.setTime(actual.getTime());
-
-        int month = expected.get(Calendar.MONTH);
-        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, "LAST");
+        int dayOfWeek = expected.get(Calendar.DAY_OF_WEEK);
+        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, dayOfWeek);
 
         filter.doActivate(componentContext);
         filter.adjustExpires(actual);
 
         assertTrue(DateUtils.isSameInstant(expected, actual));
-        assertEquals(month, actual.get(Calendar.MONTH));
+        assertEquals(dayOfWeek, actual.get(Calendar.DAY_OF_WEEK));
     }
+
+    @Test
+    public void testAdjustExpiresFutureWeekday() throws Exception {
+
+
+        Calendar actual = Calendar.getInstance();
+        actual.set(Calendar.SECOND, 0);
+        actual.set(Calendar.MILLISECOND, 0);
+
+        Calendar expected = Calendar.getInstance();
+        expected.setTime(actual.getTime());
+        expected.add(Calendar.DAY_OF_WEEK, 2);
+
+        int dayOfweek = expected.get(Calendar.DAY_OF_WEEK);
+        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, dayOfweek);
+
+
+        filter.doActivate(componentContext);
+        filter.adjustExpires(actual);
+
+        assertTrue(DateUtils.isSameInstant(expected, actual));
+        assertEquals(dayOfweek, actual.get(Calendar.DAY_OF_WEEK));
+    }
+
 
     @Test
     public void testDoActivateSuccess() throws Exception {
 
-        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, 15);
+        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, Calendar.SUNDAY);
 
         Calendar actual = Calendar.getInstance();
         actual.set(Calendar.SECOND, 0);
@@ -212,9 +170,8 @@ public class MonthlyExpiresHeaderFilterTest {
         filter.doActivate(componentContext);
         filter.adjustExpires(actual);
 
-        assertEquals(15, actual.get(Calendar.DAY_OF_MONTH));
-        // Abstract class will pass expires, this checks we aren't messing with
-        // the time.
+        assertEquals(Calendar.SUNDAY, actual.get(Calendar.DAY_OF_WEEK));
+        // Abstract class will pass expires, this checks we aren't messing with the time.
         assertNotEquals(2, actual.get(Calendar.HOUR_OF_DAY));
         assertNotEquals(30, actual.get(Calendar.MINUTE));
     }
@@ -222,27 +179,25 @@ public class MonthlyExpiresHeaderFilterTest {
     @Test(expected = ConfigurationException.class)
     public void testDoActivateCallsParent() throws Exception {
 
-        properties.remove(MonthlyExpiresHeaderFilter.PROP_EXPIRES_TIME);
+        properties.remove(WeeklyExpiresHeaderFilter.PROP_EXPIRES_TIME);
         filter.doActivate(componentContext);
     }
 
     @Test(expected = ConfigurationException.class)
-    public void testDoActivateNoDayOfMonth() throws Exception {
+    public void testDoActivateNoDayOfWeek() throws Exception {
 
         filter.doActivate(componentContext);
     }
 
     @Test(expected = ConfigurationException.class)
-    public void testDoActivateInvalidLowDayOfMonth() throws Exception {
-        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, 0);
+    public void testDoActivateInvalidLowDayOfWeek() throws Exception {
+        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, Calendar.SUNDAY - 1);
         filter.doActivate(componentContext);
     }
 
     @Test(expected = ConfigurationException.class)
-    public void testDoActivateInvalidHighDayOfMonth() throws Exception {
-        Calendar test = Calendar.getInstance();
-        int val = test.getActualMaximum(Calendar.DAY_OF_MONTH) + 1;
-        properties.put(MonthlyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_MONTH, val);
+    public void testDoActivateInvalidHighDayOfWeek() throws Exception {
+        properties.put(WeeklyExpiresHeaderFilter.PROP_EXPIRES_DAY_OF_WEEK, Calendar.SATURDAY + 1);
         filter.doActivate(componentContext);
     }
 }
